requirements.txt
jax
dm-haiku
optax
torch